[Gold V] Title: 용액, Time: 288 ms, Memory: 31624 KB -BaekjoonHub

+++ 백준/Gold/2467. 용액/용액.java
@@ -17,7 +17,7 @@ public class Main {
             liquid[i] = Integer.parseInt(st.nextToken());
         }
 
-        answer = Integer.MAX_VALUE;
+        answer = Long.MAX_VALUE;
         bs();
     }
